Less aggressive language selection
